+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 ## H19.5
 
 # TODO: 
-* (New) HAPI_Shutdown
-* (New) Session logging
 * (New) HAPI_CookPDGAllOutputs
 * HAPI_GetInstanceTransforms -> HAPI_GetInstanceTransformsOnPart
 * HAPI_GetPDGGraphContexts + HAPI_GetPDGGraphContextsCount
 * Bunch of new PDG APIs
+* Think about `Into<NodeHandle>`